refact(manager): split access content tabs into separate files
